ok moused input works, buyt now movement, working on that now

+++ src/Entity/EntityTypes/CharEntity/CoreCharEntity/CoreCharEntity.h
@@ -13,7 +13,7 @@ class CoreCharEntity : public CoreEntity
     public:
         int health;
         float speed = 0.1f;
-        CoreCharEntity(EntityRig entRig, int health , int speed, glm::vec3 pos = glm::vec3{0.0f, 0.0f, 0.0f}, glm::vec3 orient = glm::vec3{0.0f, 0.0f, 1.0f});
+        CoreCharEntity(EntityRig entRig, int health , float speed, glm::vec3 pos = glm::vec3{0.0f, 0.0f, 0.0f}, glm::vec3 orient = glm::vec3{0.0f, 0.0f, 1.0f});
         bool isDead() const {  std::cout << "player health " << health << std::endl; return health <= 0;}
         // void Draw(Shader& shader) override;
